requirements.txt
Django>=4.2,<5.0  
djangorestframework>=3.14  
drf-yasg>=1.21.4  
requests>=2.28  
pytest>=7.0  
pytest-django>=4.5  